Merge Nightly into Master

+++ src/widgets/helper/ChannelView.hpp
@@ -4,6 +4,7 @@
 #include "messages/LimitedQueue.hpp"
 #include "messages/LimitedQueueSnapshot.hpp"
 #include "messages/Selection.hpp"
+#include "messages/Image.hpp"
 #include "widgets/BaseWidget.hpp"
 
 #include <QPaintEvent>
@@ -25,7 +26,7 @@ using ChannelPtr = std::shared_ptr<Channel>;
 struct Message;
 using MessagePtr = std::shared_ptr<const Message>;
 
-enum class MessageFlag : uint16_t;
+enum class MessageFlag : uint32_t;
 using MessageFlags = FlagsEnum<MessageFlag>;
 
 class MessageLayout;
